refactor schema, create `RepoRequestBase` to minimize duplicated repo fields (#12)
