+++ api/schemas/auth.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel, Field
+
+
+class AuthorizationConfig(BaseModel):
+    code: str = Field(..., description="Authorization code")
