+++ api/schemas/repo.py
@@ -1,34 +1,16 @@
-from typing import Literal
-from urllib.parse import unquote
+from pydantic import BaseModel
 
-from pydantic import BaseModel, Field, field_validator
+from api.schemas.base import RepoRequestBase
 
 
-class RepoPrepareRequest(BaseModel):
+class RepoPrepareRequest(RepoRequestBase):
     """Request body for POST /repo/prepare (index warming). No chat messages."""
 
-    repo_url: str = Field(..., description="URL or local path of the repository")
-    type: Literal["local", "github", "gitlab", "bitbucket"] | None = Field(
-        "github", description="Repository type"
-    )
-    token: str | None = Field(None, description="PAT for private repositories")
-    provider: str = Field("google", description="Model provider (for RAG init)")
-    model: str | None = Field(None, description="Model name for the provider")
 
-    excluded_dirs: list[str] = Field(default_factory=list)
-    excluded_files: list[str] = Field(default_factory=list)
-    included_dirs: list[str] = Field(default_factory=list)
-    included_files: list[str] = Field(default_factory=list)
-
-    @field_validator(
-        "excluded_dirs",
-        "excluded_files",
-        "included_dirs",
-        "included_files",
-        mode="before",
-    )
-    @classmethod
-    def validate_path(cls, value: list[str] | str) -> list[str]:
-        if isinstance(value, str):
-            value = [unquote(p) for p in value.strip().split("\n") if p]
-        return value
+class RepoInfo(BaseModel):
+    owner: str
+    repo: str
+    type: str
+    token: str | None = None
+    localPath: str | None = None
+    repoUrl: str | None = None
